weather_app/backend/requirements.txt
fastapi
uvicorn
langchain
langchain-community
langchain-openai
langgraph
requests
python-dotenv